open-r1-multimodal/requirements_current.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
bitsandbytes==0.45.3
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.3.2
deepspeed==0.15.4
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
filelock==3.17.0
flake8==7.1.2
flash_attn==2.7.4.post1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
hf_transfer==0.1.9
hjson==3.1.0
huggingface-hub==0.29.1
idna==3.10
iniconfig==2.0.0
inquirerpy==0.3.4
isort==6.0.1
Jinja2==3.1.5
latex2sympy2_extended==1.10.1
liger_kernel==0.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.7.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
-e git+https://github.com/om-ai-lab/VLM-R1.git@fc92ce8d9221a566b80fe89f5adc6b13434a76ad#egg=open_r1&subdirectory=src/open-r1-multimodal
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboardX==2.6.2.2
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@02776d2c6aa997c5b81f28f2edf38df9967253be
triton==3.2.0
trl @ git+https://github.com/huggingface/trl.git@b882f57d9387c224371e6c827c2e9b0786e9aed3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.18.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
